--- src/main/java/com/colorit/backend/storages/AbstractStorageResponse.java
@@ -1,4 +1,0 @@
-package com.colorit.backend.storages;
-
-public class AbstractStorageResponse {
-}
