+++ src/main/java/com/colorit/backend/storages/FileStorage.java
@@ -1,11 +1,15 @@
 package com.colorit.backend.storages;
 
+import com.colorit.backend.storages.responses.AbstractStorageResponse;
+import com.colorit.backend.storages.storageimpls.CloudinaryStorage;
+import com.colorit.backend.storages.storageimpls.LocalStorage;
+
 import java.util.Map;
 
 public class FileStorage implements IStorage {
     private IStorage storage;
     private final static String HEROKU_VAR = "HEROKU_VAR";
-    private final static String OS_NAME = System.getProperty("os.name");
+    private final static String USER_HOME = System.getProperty("os.name");
 
     // TODO create storage, by detecting env_var -> heroku => then create Cloudinary, else local
     FileStorage() {
@@ -13,7 +17,8 @@ public class FileStorage implements IStorage {
         if (envVars.get(HEROKU_VAR) != null) {
             this.storage = new CloudinaryStorage();
         } else {
-            this.storage = new LocalStorage();
+
+            this.storage = new LocalStorage(USER_HOME);
         }
     }
