+++ src/main/java/com/colorit/backend/storages/IStorage.java
@@ -1,5 +1,7 @@
 package com.colorit.backend.storages;
 
+import com.colorit.backend.storages.responses.AbstractStorageResponse;
+
 public interface IStorage {
     AbstractStorageResponse writeFile();
