--- src/main/java/com/colorit/backend/storages/LocalStorage.java
@@ -1,13 +1,0 @@
-package com.colorit.backend.storages;
-
-public class LocalStorage implements IStorage {
-    @Override
-    public AbstractStorageResponse writeFile() {
-        return null;
-    }
-
-    @Override
-    public AbstractStorageResponse readFile() {
-        return null;
-    }
-}
