+++ src/main/java/com/colorit/backend/storages/responses/AbstractStorageResponse.java
@@ -1,0 +1,4 @@
+package com.colorit.backend.storages.responses;
+
+public class AbstractStorageResponse {
+}
