+++ src/main/resources/application.properties
@@ -7,3 +7,6 @@ spring.datasource.hikari.maximum-pool-size=1
 spring.jpa.hibernate.ddl-auto=update
 #spring.http.multipart.max-file-size=128KB
 #spring.http.multipart.max-request-size=128KB
+api_key=API_KEY
+api_secret=API_SECRET
+cloud_name=CLOUD_NAME
